pyproject.toml
[tool.poetry]
name = "w01-projeto-processo-de-dados-do-zero"
version = "0.1.0"
description = ""
authors = ["Matheus Costa <matheeusferreeira@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.12.3"
streamlit = "^1.33.0"
pydantic = {extras = ["email"], version = "^2.7.0"}
pytest = "^8.1.1"
selenium = "^4.19.0"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
isort = "^5.13.2"
taskipy = "^1.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "streamlit run src/app.py"
test = "pytest tests -v"
